// src/types/buttonType.tsx
import { ReactElement } from "react";
export interface ButtonProps {
  title?: any;
  backgroundColor?: any;
  color?: any;
  borderRadius?: any;
  startIcon?: any;
  padding?: any;
  boxShadow?: any;
  width?: any,
  maxWidth?: any;
  textTransform?: any;
  fontSize?: any;
  fontWeight?: any;
  lineHeight?: any;
  endIcon?: ReactElement;
  margin?: any;
  height?: any;
  border?: any;
  whiteSpace?: any;
  letterSpacing?: any;
  className?: any;
  disabled?: any;
  onClick?: () => void;
}